backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
numpy>=1.24.0
nvidia-riva-client>=2.14.0
